The first look
Created the first page when a teacher looks into the page.

Move from the first task a bit. Even tho it asks so that the first look into the page would only allow to create the project, I made it a bit more universal. Every time teacher goes to the page, it would allow to see all created projects or create a new one. This will allow for less work with better usability.

Design is simple and easy to understand. A few css lines to create a form and tables.

.htaccess will allow to control everything in index file with nice and elegant /name url.