System redesign
I had a dream where I argued with myself on how the best way to implement mysql with current objects, and I think I solved it.

Now mysql is implemented in the Project class, so that it will create itself by the given ID in the contructor, add everything where it needs, and I can control everything with the Project class. Simple and clean